Update and rename README.md to Tic tac toe bot (code soft project)
This Python project is a straightforward implementation of the classic Tic Tac Toe game. It provides a console-based interface for two players to engage in the timeless battle of Xs and Os. The code is designed to be clear and concise, making it an excellent resource for beginners looking to understand the fundamentals of programming logic and game development